extract vm template link to a file
