+++ frontend/packages/kubevirt-plugin/src/components/vm-templates/vm-template-link.tsx
@@ -1,0 +1,29 @@
+import * as React from 'react';
+import { Link } from 'react-router-dom';
+
+import { ResourceIcon } from '@console/internal/components/utils';
+import { getName, getNamespace, getUid } from '@console/shared';
+import { TemplateModel } from '@console/internal/models';
+import { TemplateKind } from '@console/internal/module/k8s';
+
+export const VMTemplateLink: React.FC<VMTemplateLinkProps> = ({ template }) => {
+  const name = getName(template);
+  const namespace = getNamespace(template);
+
+  return (
+    <>
+      <ResourceIcon kind={TemplateModel.kind} />
+      <Link
+        to={`/k8s/ns/${namespace}/vmtemplates/${name}`}
+        title={getUid(template)}
+        className="co-resource-item__resource-name"
+      >
+        {name}
+      </Link>
+    </>
+  );
+};
+
+type VMTemplateLinkProps = {
+  template: TemplateKind;
+};
